requirements.txt
pygame-ce>=2.5.0
pymunk>=6.6.0
pytest>=8.0
